pyproject.toml
[project]
name = "stream"
version = "0.1.0"
description = "Groundlight Stream Processor - Container for analyzing video using RTSP etc"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "framegrab==0.7.0",
    "groundlight==0.18.4",
    "docopt==0.6.2",
    "jsonformatter==0.3.2",
    "numpy<2.0.0",
    "opencv-python-headless==4.8.1.78",
    "PyYAML==6.0.2",
    "youtube-dl==2021.12.17",
    "pafy==0.5.5"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff==0.7.2",
]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]

[tool.ruff]
exclude = ["__init__.py"]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "PL", "UP"] # https://beta.ruff.rs/docs/rules/
ignore = [
    "E501",  # Line length too long (handled by formatter)
]
fixable = ["ALL"]  # Allow autofix behavior
unfixable = []

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"